Move `DeltaMap` type to separate module
