+++ lib/dbvar/dbvar.cabal
@@ -50,6 +50,7 @@ library
       Data.Chain
       Data.DBVar
       Data.Delta
+      Data.DeltaMap
       Data.Table
       Database.Persist.Delta
       Database.Schema
